progress on find wall functionality

+++ src/turtlebot3_simulations/vroom/include/odom.hpp
@@ -14,7 +14,7 @@ This should then store this in a member called robot_pose_
 #include <tf2/LinearMath/Matrix3x3.h>
 #include <tf2/LinearMath/Quaternion.h>
 #include <rclcpp/rclcpp.hpp>
-#include "std_msgs/msg/float64.hpp"
+#include "std_msgs/msg/float32.hpp"
 
 
 class Odom : public rclcpp::Node
@@ -32,7 +32,7 @@ class Odom : public rclcpp::Node
         void odom_callback(const nav_msgs::msg::Odometry::SharedPtr msg);
 
         // Odometry Publisher
-        rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr odom_pub_; 
+        rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr odom_pub_; 
         rclcpp::TimerBase::SharedPtr update_timer_;
         void update_pose();
